producer/requirements.txt
kafka-python
lz4
